rename : screen-record/record 네이밍 변경
