+++ client/src/features/interview/screen-record/index.ts
@@ -1,1 +1,2 @@
 export { default as InterviewRecord } from './ui/InterviewRecord';
+export { default as InterviewRecordSettings } from './ui/setting/InterviewRecordSetting';
