+++ client/src/features/interview/screen-record/ui/record/index.ts
@@ -1,3 +1,3 @@
-export { default as VideoPlayer } from './VideoPlayer';
+export { default as InterviewRecordingQueue } from './InterviewRecordingQueue ';
 export { default as ScriptTextArea } from './ScriptTextArea';
 export { default as RecordActionButtons } from './RecordActionButtons';
