--- client/src/features/interview/screen-record/ui/setting/index.ts
@@ -1,1 +1,0 @@
-export { default as InterviewRecordSettings } from './InterviewRecordSetting';
